requirements.txt
flask
pillow
numpy
torch
onnx
onnxruntime
opencv-python-headless
pandas
requests